+++ arguments-builder.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "@iringo/arguments-builder";
-import { api, logLevel, nextHour, output, storage, weather } from "./arguments-builder-full.config";
+import { airQuality, api, logLevel, nextHour, output, storage, weather } from "./arguments-builder-full.config";
 
 export default defineConfig({
     output: output,
-    args: [...weather, ...nextHour, ...api, ...storage, ...logLevel],
+    args: [...weather, ...nextHour, ...airQuality, ...api, ...storage, ...logLevel],
 });