+++ src/function/setENV.mjs
@@ -15,6 +15,8 @@ export default function setENV(name, platforms, database) {
     /***************** Settings *****************/
     Console.info(`typeof Settings: ${typeof Settings}`, `Settings: ${JSON.stringify(Settings, null, 2)}`);
     string2array("Weather.Replace");
+    string2array("AirQuality.Replace");
+    string2array("AirQuality.Fallback.Provider");
     /***************** Caches *****************/
     //Console.debug(`typeof Caches: ${typeof Caches}`, `Caches: ${JSON.stringify(Caches)}`);
     /***************** Configs *****************/